Fully extract more sync logic into the VaultSyncManager
