+++ app/src/main/kotlin/com/x8bit/bitwarden/data/vault/manager/model/ImportCxfPayloadResult.kt
@@ -16,6 +16,11 @@ sealed class ImportCxfPayloadResult {
     data object NoItems : ImportCxfPayloadResult()
 
     /**
+     * The sync process has failed after importing the CXF payload.
+     */
+    data class SyncFailed(val error: Throwable) : ImportCxfPayloadResult()
+
+    /**
      * There was an error importing the vault data.
      *
      * @param error The error that occurred during import.
