+++ app/src/test/kotlin/com/x8bit/bitwarden/data/vault/repository/VaultRepositoryTest.kt
@@ -1,13 +1,10 @@
 package com.x8bit.bitwarden.data.vault.repository
 
-import android.util.Base64
 import app.cash.turbine.test
-import app.cash.turbine.turbineScope
 import com.bitwarden.collections.CollectionView
 import com.bitwarden.core.DateTime
 import com.bitwarden.core.InitUserCryptoMethod
 import com.bitwarden.core.data.repository.model.DataState
-import com.bitwarden.core.data.repository.util.bufferedMutableSharedFlow
 import com.bitwarden.core.data.util.asFailure
 import com.bitwarden.core.data.util.asSuccess
 import com.bitwarden.data.datasource.disk.base.FakeDispatcherManager
@@ -17,11 +14,8 @@ import com.bitwarden.fido.Fido2CredentialAutofillView
 import com.bitwarden.network.model.CipherTypeJson
 import com.bitwarden.network.model.SyncResponseJson
 import com.bitwarden.network.model.createMockCipher
-import com.bitwarden.network.model.createMockCollection
-import com.bitwarden.network.model.createMockDomains
 import com.bitwarden.network.model.createMockFolder
 import com.bitwarden.network.model.createMockOrganizationKeys
-import com.bitwarden.network.model.createMockSend
 import com.bitwarden.sdk.Fido2CredentialStore
 import com.bitwarden.send.SendView
 import com.bitwarden.vault.CipherType
@@ -34,21 +28,12 @@ import com.x8bit.bitwarden.data.auth.datasource.disk.model.AccountTokensJson
 import com.x8bit.bitwarden.data.auth.datasource.disk.model.UserStateJson
 import com.x8bit.bitwarden.data.auth.datasource.disk.util.FakeAuthDiskSource
 import com.x8bit.bitwarden.data.auth.repository.util.toSdkParams
-import com.x8bit.bitwarden.data.platform.datasource.disk.SettingsDiskSource
 import com.x8bit.bitwarden.data.platform.error.MissingPropertyException
 import com.x8bit.bitwarden.data.platform.error.NoActiveUserException
-import com.x8bit.bitwarden.data.platform.manager.DatabaseSchemeManager
-import com.x8bit.bitwarden.data.platform.manager.PushManager
 import com.x8bit.bitwarden.data.vault.datasource.disk.VaultDiskSource
 import com.x8bit.bitwarden.data.vault.datasource.sdk.VaultSdkSource
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockAccount
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockCipherListView
-import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockCipherView
-import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockCollectionView
-import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockDecryptCipherListResult
-import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockFolderView
-import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkCipher
-import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkCollection
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkFolder
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkSend
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSendView
@@ -65,127 +50,79 @@ import com.x8bit.bitwarden.data.vault.repository.model.GenerateTotpResult
 import com.x8bit.bitwarden.data.vault.repository.model.ImportCredentialsResult
 import com.x8bit.bitwarden.data.vault.repository.model.SendData
 import com.x8bit.bitwarden.data.vault.repository.model.VaultData
-import com.x8bit.bitwarden.data.vault.repository.model.VaultUnlockData
 import com.x8bit.bitwarden.data.vault.repository.model.VaultUnlockResult
-import com.x8bit.bitwarden.data.vault.repository.model.createMockDomainsData
-import com.x8bit.bitwarden.data.vault.repository.util.toDomainsData
 import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkCipher
-import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkCipherList
-import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkCollectionList
-import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkFolderList
-import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkSendList
 import com.x8bit.bitwarden.ui.vault.feature.verificationcode.util.createVerificationCodeItem
-import io.mockk.awaits
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
 import io.mockk.mockkConstructor
-import io.mockk.mockkStatic
 import io.mockk.runs
 import io.mockk.unmockkConstructor
-import io.mockk.unmockkStatic
-import io.mockk.verify
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.net.UnknownHostException
 import java.security.GeneralSecurityException
-import java.security.MessageDigest
-import java.time.Clock
-import java.time.Instant
-import java.time.ZoneOffset
 import java.time.ZonedDateTime
-import java.time.temporal.ChronoUnit
 import javax.crypto.BadPaddingException
 import javax.crypto.Cipher
 
 @Suppress("LargeClass")
 class VaultRepositoryTest {
-    private val clock: Clock = Clock.fixed(
-        Instant.parse("2023-10-27T12:00:00Z"),
-        ZoneOffset.UTC,
-    )
     private val dispatcherManager: DispatcherManager = FakeDispatcherManager()
     private val fakeAuthDiskSource = FakeAuthDiskSource()
-    private val settingsDiskSource = mockk<SettingsDiskSource> {
-        every { getLastSyncTime(userId = any()) } returns clock.instant()
-    }
     private val mutableGetCiphersFlow: MutableStateFlow<List<SyncResponseJson.Cipher>> =
         MutableStateFlow(listOf(createMockCipher(1)))
     private val vaultDiskSource: VaultDiskSource = mockk {
-        coEvery { resyncVaultData(any()) } just runs
         every { getCiphersFlow(any()) } returns mutableGetCiphersFlow
     }
     private val totpCodeManager: TotpCodeManager = mockk()
-    private val vaultSdkSource: VaultSdkSource = mockk {
-        every { clearCrypto(userId = any()) } just runs
+    private val vaultSdkSource: VaultSdkSource = mockk()
+    private val vaultLockManager: VaultLockManager = mockk()
+    private val mutableVaultDataStateFlow =
+        MutableStateFlow<DataState<VaultData>>(DataState.Loading)
+    private val mutableCollectionsStateFlow =
+        MutableStateFlow<DataState<List<CollectionView>>>(DataState.Loading)
+    private val mutableDecryptCipherListResultStateFlow =
+        MutableStateFlow<DataState<DecryptCipherListResult>>(DataState.Loading)
+    private val mutableDomainsStateFlow =
+        MutableStateFlow<DataState<DomainsData>>(DataState.Loading)
+    private val mutableFoldersStateFlow =
+        MutableStateFlow<DataState<List<FolderView>>>(DataState.Loading)
+    private val mutableSendDataStateFlow = MutableStateFlow<DataState<SendData>>(DataState.Loading)
+    private val vaultSyncManager: VaultSyncManager = mockk {
+        every { vaultDataStateFlow } returns mutableVaultDataStateFlow
+        every { collectionsStateFlow } returns mutableCollectionsStateFlow
+        every { decryptCipherListResultStateFlow } returns mutableDecryptCipherListResultStateFlow
+        every { domainsStateFlow } returns mutableDomainsStateFlow
+        every { foldersStateFlow } returns mutableFoldersStateFlow
+        every { sendDataStateFlow } returns mutableSendDataStateFlow
     }
-    private val mutableVaultStateFlow = MutableStateFlow<List<VaultUnlockData>>(
-        emptyList(),
-    )
-    private val mutableUnlockedUserIdsStateFlow = MutableStateFlow<Set<String>>(emptySet())
-    private val vaultLockManager: VaultLockManager = mockk {
-        every { vaultUnlockDataStateFlow } returns mutableVaultStateFlow
-        every {
-            isVaultUnlocked(any())
-        } answers { call ->
-            val userId = call.invocation.args.first()
-            userId in mutableUnlockedUserIdsStateFlow.value
-        }
-        every { isVaultUnlocking(any()) } returns false
-        every { lockVault(any(), any()) } just runs
-        every { lockVaultForCurrentUser(any()) } just runs
-        coEvery {
-            waitUntilUnlocked(any())
-        } coAnswers { call ->
-            val userId = call.invocation.args.first()
-            mutableUnlockedUserIdsStateFlow.first { userId in it }
-        }
-    }
-    private val mutableDatabaseSchemeChangeFlow = bufferedMutableSharedFlow<Unit>()
-    private val databaseSchemeManager: DatabaseSchemeManager = mockk {
-        every { databaseSchemeChangeFlow } returns mutableDatabaseSchemeChangeFlow
-    }
-    private val mutableFullSyncFlow = bufferedMutableSharedFlow<Unit>()
-    private val pushManager: PushManager = mockk {
-        every { fullSyncFlow } returns mutableFullSyncFlow
-    }
-    private val vaultSyncManager: VaultSyncManager = mockk()
     private val credentialExchangeImportManager: CredentialExchangeImportManager = mockk()
 
     private val vaultRepository = VaultRepositoryImpl(
         vaultDiskSource = vaultDiskSource,
         vaultSdkSource = vaultSdkSource,
         authDiskSource = fakeAuthDiskSource,
-        settingsDiskSource = settingsDiskSource,
         vaultLockManager = vaultLockManager,
         dispatcherManager = dispatcherManager,
         totpCodeManager = totpCodeManager,
-        pushManager = pushManager,
         cipherManager = mockk(),
         folderManager = mockk(),
         sendManager = mockk(),
-        clock = clock,
-        databaseSchemeManager = databaseSchemeManager,
         vaultSyncManager = vaultSyncManager,
         credentialExchangeImportManager = credentialExchangeImportManager,
     )
 
     @BeforeEach
     fun setup() {
-        mockkStatic(SyncResponseJson.Domains::toDomainsData)
-        mockkStatic(MessageDigest::class)
-        mockkStatic(Base64::class)
         mockkConstructor(NoActiveUserException::class)
         mockkConstructor(MissingPropertyException::class)
         every {
@@ -198,531 +135,8 @@ class VaultRepositoryTest {
 
     @AfterEach
     fun tearDown() {
-        unmockkStatic(SyncResponseJson.Domains::toDomainsData)
-        unmockkStatic(MessageDigest::class)
-        unmockkStatic(Base64::class)
         unmockkConstructor(NoActiveUserException::class)
         unmockkConstructor(MissingPropertyException::class)
-    }
-
-    @Test
-    fun `userSwitchingChangesFlow should cancel any pending sync call`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        coEvery { vaultSyncManager.sync(any(), any()) } just awaits
-
-        vaultRepository.sync()
-        vaultRepository.sync()
-        coVerify(exactly = 1) {
-            // Despite being called twice, we only allow 1 sync
-            vaultSyncManager.sync(any(), any())
-        }
-
-        fakeAuthDiskSource.userState = UserStateJson(
-            activeUserId = "mockId-2",
-            accounts = mapOf("mockId-2" to mockk()),
-        )
-        vaultRepository.sync()
-        coVerify {
-            // A second sync should have happened now since it was cancelled by the userState change
-            vaultSyncManager.sync(userId = "mockId-1", forced = any())
-            vaultSyncManager.sync(userId = "mockId-2", forced = any())
-        }
-    }
-
-    @Test
-    fun `userSwitchingChangesFlow should clear unlocked data`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        val userId = "mockId-1"
-        coEvery {
-            vaultSdkSource.decryptCipherListWithFailures(
-                userId = userId,
-                cipherList = listOf(createMockSdkCipher(1, clock)),
-            )
-        } returns createMockDecryptCipherListResult(number = 1).asSuccess()
-        coEvery {
-            vaultSdkSource.decryptFolderList(
-                userId = userId,
-                folderList = listOf(createMockSdkFolder(1)),
-            )
-        } returns listOf(createMockFolderView(number = 1)).asSuccess()
-        coEvery {
-            vaultSdkSource.decryptCollectionList(
-                userId = userId,
-                collectionList = listOf(createMockSdkCollection(1)),
-            )
-        } returns listOf(createMockCollectionView(number = 1)).asSuccess()
-        coEvery {
-            vaultSdkSource.decryptSendList(
-                userId = userId,
-                sendList = listOf(createMockSdkSend(number = 1)),
-            )
-        } returns listOf(createMockSendView(number = 1)).asSuccess()
-        val ciphersFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Cipher>>()
-        val collectionsFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Collection>>()
-        val foldersFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Folder>>()
-        val sendsFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Send>>()
-        val domainsFlow = bufferedMutableSharedFlow<SyncResponseJson.Domains>()
-        setupVaultDiskSourceFlows(
-            ciphersFlow = ciphersFlow,
-            collectionsFlow = collectionsFlow,
-            foldersFlow = foldersFlow,
-            sendsFlow = sendsFlow,
-            domainsFlow = domainsFlow,
-        )
-
-        turbineScope {
-            val ciphersStateFlow =
-                vaultRepository.decryptCipherListResultStateFlow.testIn(backgroundScope)
-            val collectionsStateFlow = vaultRepository.collectionsStateFlow.testIn(backgroundScope)
-            val foldersStateFlow = vaultRepository.foldersStateFlow.testIn(backgroundScope)
-            val sendsStateFlow = vaultRepository.sendDataStateFlow.testIn(backgroundScope)
-            val domainsStateFlow = vaultRepository.domainsStateFlow.testIn(backgroundScope)
-
-            assertEquals(DataState.Loading, ciphersStateFlow.awaitItem())
-            assertEquals(DataState.Loading, collectionsStateFlow.awaitItem())
-            assertEquals(DataState.Loading, foldersStateFlow.awaitItem())
-            assertEquals(DataState.Loading, sendsStateFlow.awaitItem())
-            assertEquals(DataState.Loading, domainsStateFlow.awaitItem())
-
-            ciphersFlow.tryEmit(listOf(createMockCipher(number = 1)))
-            collectionsFlow.tryEmit(listOf(createMockCollection(number = 1)))
-            foldersFlow.tryEmit(listOf(createMockFolder(number = 1)))
-            sendsFlow.tryEmit(listOf(createMockSend(number = 1)))
-            domainsFlow.tryEmit(createMockDomains(number = 1))
-
-            // No events received until unlocked
-            ciphersStateFlow.expectNoEvents()
-            collectionsStateFlow.expectNoEvents()
-            foldersStateFlow.expectNoEvents()
-            sendsStateFlow.expectNoEvents()
-            // Domains does not care about being unlocked
-            assertEquals(
-                DataState.Loaded(createMockDomainsData(number = 1)),
-                domainsStateFlow.awaitItem(),
-            )
-
-            setVaultToUnlocked(userId = userId)
-
-            ciphersFlow.tryEmit(listOf(createMockCipher(number = 1)))
-            collectionsFlow.tryEmit(listOf(createMockCollection(number = 1)))
-            foldersFlow.tryEmit(listOf(createMockFolder(number = 1)))
-            sendsFlow.tryEmit(listOf(createMockSend(number = 1)))
-            domainsFlow.tryEmit(createMockDomains(number = 1))
-
-            assertEquals(
-                DataState.Loaded(createMockDecryptCipherListResult(number = 1)),
-                ciphersStateFlow.awaitItem(),
-            )
-            assertEquals(
-                DataState.Loaded(listOf(createMockCollectionView(number = 1))),
-                collectionsStateFlow.awaitItem(),
-            )
-            assertEquals(
-                DataState.Loaded(listOf(createMockFolderView(number = 1))),
-                foldersStateFlow.awaitItem(),
-            )
-            assertEquals(
-                DataState.Loaded(SendData(listOf(createMockSendView(number = 1)))),
-                sendsStateFlow.awaitItem(),
-            )
-            // Domain data has not changed
-            domainsStateFlow.expectNoEvents()
-
-            fakeAuthDiskSource.userState = null
-
-            assertEquals(DataState.Loading, ciphersStateFlow.awaitItem())
-            assertEquals(DataState.Loading, collectionsStateFlow.awaitItem())
-            assertEquals(DataState.Loading, foldersStateFlow.awaitItem())
-            assertEquals(DataState.Loading, sendsStateFlow.awaitItem())
-            assertEquals(DataState.Loading, domainsStateFlow.awaitItem())
-        }
-    }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `mutableVaultStateFlow should clear unlocked data when it does not contain the active user`() =
-        runTest {
-            fakeAuthDiskSource.userState = UserStateJson(
-                activeUserId = "mockId-1",
-                accounts = mapOf(
-                    "mockId-1" to MOCK_ACCOUNT,
-                    "mockId-2" to mockk(),
-                ),
-            )
-            mutableVaultStateFlow.value = listOf(
-                VaultUnlockData(userId = "mockId-1", status = VaultUnlockData.Status.UNLOCKING),
-                VaultUnlockData(userId = "mockId-2", status = VaultUnlockData.Status.UNLOCKED),
-            )
-            val userId = "mockId-1"
-            coEvery {
-                vaultSdkSource.decryptCipherListWithFailures(
-                    userId = userId,
-                    cipherList = listOf(createMockSdkCipher(1, clock)),
-                )
-            } returns createMockDecryptCipherListResult(number = 1).asSuccess()
-            coEvery {
-                vaultSdkSource.decryptFolderList(
-                    userId = userId,
-                    folderList = listOf(createMockSdkFolder(1)),
-                )
-            } returns listOf(createMockFolderView(number = 1)).asSuccess()
-            coEvery {
-                vaultSdkSource.decryptCollectionList(
-                    userId = userId,
-                    collectionList = listOf(createMockSdkCollection(1)),
-                )
-            } returns listOf(createMockCollectionView(number = 1)).asSuccess()
-            coEvery {
-                vaultSdkSource.decryptSendList(
-                    userId = userId,
-                    sendList = listOf(createMockSdkSend(number = 1)),
-                )
-            } returns listOf(createMockSendView(number = 1)).asSuccess()
-            val ciphersFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Cipher>>()
-            val collectionsFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Collection>>()
-            val foldersFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Folder>>()
-            val sendsFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Send>>()
-            val domainsFlow = bufferedMutableSharedFlow<SyncResponseJson.Domains>()
-            setupVaultDiskSourceFlows(
-                ciphersFlow = ciphersFlow,
-                collectionsFlow = collectionsFlow,
-                foldersFlow = foldersFlow,
-                sendsFlow = sendsFlow,
-                domainsFlow = domainsFlow,
-            )
-
-            turbineScope {
-                val ciphersStateFlow =
-                    vaultRepository.decryptCipherListResultStateFlow.testIn(backgroundScope)
-                val collectionsStateFlow =
-                    vaultRepository.collectionsStateFlow.testIn(backgroundScope)
-                val foldersStateFlow = vaultRepository.foldersStateFlow.testIn(backgroundScope)
-                val sendsStateFlow = vaultRepository.sendDataStateFlow.testIn(backgroundScope)
-                val domainsStateFlow = vaultRepository.domainsStateFlow.testIn(backgroundScope)
-
-                assertEquals(DataState.Loading, ciphersStateFlow.awaitItem())
-                assertEquals(DataState.Loading, collectionsStateFlow.awaitItem())
-                assertEquals(DataState.Loading, foldersStateFlow.awaitItem())
-                assertEquals(DataState.Loading, sendsStateFlow.awaitItem())
-                assertEquals(DataState.Loading, domainsStateFlow.awaitItem())
-
-                ciphersFlow.tryEmit(listOf(createMockCipher(number = 1)))
-                collectionsFlow.tryEmit(listOf(createMockCollection(number = 1)))
-                foldersFlow.tryEmit(listOf(createMockFolder(number = 1)))
-                sendsFlow.tryEmit(listOf(createMockSend(number = 1)))
-                domainsFlow.tryEmit(createMockDomains(number = 1))
-
-                // No events received until unlocked
-                ciphersStateFlow.expectNoEvents()
-                collectionsStateFlow.expectNoEvents()
-                foldersStateFlow.expectNoEvents()
-                sendsStateFlow.expectNoEvents()
-                // Domains does not care about being unlocked
-                assertEquals(
-                    DataState.Loaded(createMockDomainsData(number = 1)),
-                    domainsStateFlow.awaitItem(),
-                )
-                setVaultToUnlocked(userId = userId)
-
-                assertEquals(
-                    DataState.Loaded(createMockDecryptCipherListResult(number = 1)),
-                    ciphersStateFlow.awaitItem(),
-                )
-                assertEquals(
-                    DataState.Loaded(listOf(createMockCollectionView(number = 1))),
-                    collectionsStateFlow.awaitItem(),
-                )
-                assertEquals(
-                    DataState.Loaded(listOf(createMockFolderView(number = 1))),
-                    foldersStateFlow.awaitItem(),
-                )
-                assertEquals(
-                    DataState.Loaded(SendData(listOf(createMockSendView(number = 1)))),
-                    sendsStateFlow.awaitItem(),
-                )
-                // Domain data has not changed
-                domainsStateFlow.expectNoEvents()
-
-                mutableVaultStateFlow.value = listOf(
-                    VaultUnlockData(userId = "mockId-2", status = VaultUnlockData.Status.UNLOCKED),
-                )
-
-                assertEquals(DataState.Loading, ciphersStateFlow.awaitItem())
-                assertEquals(DataState.Loading, collectionsStateFlow.awaitItem())
-                assertEquals(DataState.Loading, foldersStateFlow.awaitItem())
-                assertEquals(DataState.Loading, sendsStateFlow.awaitItem())
-                domainsStateFlow.expectNoEvents()
-            }
-        }
-
-    @Test
-    fun `ciphersStateFlow should emit decrypted list of ciphers when decryptCipherList succeeds`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockCipherList = listOf(createMockCipher(number = 1))
-            val mockEncryptedCipherList = mockCipherList.toEncryptedSdkCipherList()
-            val mockDecryptCipherListResult = createMockDecryptCipherListResult(number = 1)
-            val mutableCiphersStateFlow =
-                bufferedMutableSharedFlow<List<SyncResponseJson.Cipher>>(replay = 1)
-            every {
-                vaultDiskSource.getCiphersFlow(userId = MOCK_USER_STATE.activeUserId)
-            } returns mutableCiphersStateFlow
-            coEvery {
-                vaultSdkSource.decryptCipherListWithFailures(
-                    userId = userId,
-                    cipherList = mockEncryptedCipherList,
-                )
-            } returns mockDecryptCipherListResult.asSuccess()
-
-            vaultRepository
-                .decryptCipherListResultStateFlow
-                .test {
-                    assertEquals(DataState.Loading, awaitItem())
-                    mutableCiphersStateFlow.tryEmit(mockCipherList)
-
-                    // No additional emissions until vault is unlocked
-                    expectNoEvents()
-                    setVaultToUnlocked(userId = userId)
-
-                    assertEquals(DataState.Loaded(mockDecryptCipherListResult), awaitItem())
-                }
-        }
-
-    @Test
-    fun `ciphersStateFlow should emit an error when decryptCipherList fails`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        val userId = "mockId-1"
-        val throwable = Throwable("Fail")
-        val mockCipherList = listOf(createMockCipher(number = 1))
-        val mockEncryptedCipherList = mockCipherList.toEncryptedSdkCipherList()
-        val mutableCiphersStateFlow =
-            bufferedMutableSharedFlow<List<SyncResponseJson.Cipher>>(replay = 1)
-        every {
-            vaultDiskSource.getCiphersFlow(userId = MOCK_USER_STATE.activeUserId)
-        } returns mutableCiphersStateFlow
-        coEvery {
-            vaultSdkSource.decryptCipherListWithFailures(
-                userId = userId,
-                cipherList = mockEncryptedCipherList,
-            )
-        } returns throwable.asFailure()
-
-        vaultRepository
-            .decryptCipherListResultStateFlow
-            .test {
-                assertEquals(DataState.Loading, awaitItem())
-                mutableCiphersStateFlow.tryEmit(mockCipherList)
-
-                // No additional emissions until vault is unlocked
-                expectNoEvents()
-                setVaultToUnlocked(userId = userId)
-
-                assertEquals(DataState.Error<DecryptCipherListResult>(throwable), awaitItem())
-            }
-    }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `collectionsStateFlow should emit decrypted list of collections when decryptCollectionList succeeds`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockCollectionList = listOf(createMockCollection(number = 1))
-            val mockEncryptedCollectionList = mockCollectionList.toEncryptedSdkCollectionList()
-            val mockCollectionViewList = listOf(createMockCollectionView(number = 1))
-            val mutableCollectionsStateFlow =
-                bufferedMutableSharedFlow<List<SyncResponseJson.Collection>>(replay = 1)
-            every {
-                vaultDiskSource.getCollections(userId = MOCK_USER_STATE.activeUserId)
-            } returns mutableCollectionsStateFlow
-            coEvery {
-                vaultSdkSource.decryptCollectionList(
-                    userId = userId,
-                    collectionList = mockEncryptedCollectionList,
-                )
-            } returns mockCollectionViewList.asSuccess()
-
-            vaultRepository
-                .collectionsStateFlow
-                .test {
-                    assertEquals(DataState.Loading, awaitItem())
-                    mutableCollectionsStateFlow.tryEmit(mockCollectionList)
-
-                    // No additional emissions until vault is unlocked
-                    expectNoEvents()
-                    setVaultToUnlocked(userId = userId)
-
-                    assertEquals(DataState.Loaded(mockCollectionViewList), awaitItem())
-                }
-        }
-
-    @Test
-    fun `collectionsStateFlow should emit an error when decryptCollectionList fails`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        val userId = "mockId-1"
-        val throwable = Throwable("Fail")
-        val mockCollectionList = listOf(createMockCollection(number = 1))
-        val mockEncryptedCollectionList = mockCollectionList.toEncryptedSdkCollectionList()
-        val mutableCollectionStateFlow =
-            bufferedMutableSharedFlow<List<SyncResponseJson.Collection>>(replay = 1)
-        every {
-            vaultDiskSource.getCollections(userId = MOCK_USER_STATE.activeUserId)
-        } returns mutableCollectionStateFlow
-        coEvery {
-            vaultSdkSource.decryptCollectionList(
-                userId = userId,
-                collectionList = mockEncryptedCollectionList,
-            )
-        } returns throwable.asFailure()
-
-        vaultRepository
-            .collectionsStateFlow
-            .test {
-                assertEquals(DataState.Loading, awaitItem())
-                mutableCollectionStateFlow.tryEmit(mockCollectionList)
-
-                // No additional emissions until vault is unlocked
-                expectNoEvents()
-                setVaultToUnlocked(userId = userId)
-
-                assertEquals(DataState.Error<List<CollectionView>>(throwable), awaitItem())
-            }
-    }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `foldersStateFlow should emit decrypted list of folders when decryptFolderList succeeds`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockFolderList = listOf(createMockFolder(number = 1))
-            val mockEncryptedFolderList = mockFolderList.toEncryptedSdkFolderList()
-            val mockFolderViewList = listOf(createMockFolderView(number = 1))
-            val mutableFoldersStateFlow =
-                bufferedMutableSharedFlow<List<SyncResponseJson.Folder>>(replay = 1)
-            every {
-                vaultDiskSource.getFolders(userId = MOCK_USER_STATE.activeUserId)
-            } returns mutableFoldersStateFlow
-            coEvery {
-                vaultSdkSource.decryptFolderList(
-                    userId = userId,
-                    folderList = mockEncryptedFolderList,
-                )
-            } returns mockFolderViewList.asSuccess()
-
-            vaultRepository
-                .foldersStateFlow
-                .test {
-                    assertEquals(DataState.Loading, awaitItem())
-                    mutableFoldersStateFlow.tryEmit(mockFolderList)
-
-                    // No additional emissions until vault is unlocked
-                    expectNoEvents()
-                    setVaultToUnlocked(userId = userId)
-
-                    assertEquals(DataState.Loaded(mockFolderViewList), awaitItem())
-                }
-        }
-
-    @Test
-    fun `foldersStateFlow should emit an error when decryptFolderList fails`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        val userId = "mockId-1"
-        val throwable = Throwable("Fail")
-        val mockFolderList = listOf(createMockFolder(number = 1))
-        val mockEncryptedFolderList = mockFolderList.toEncryptedSdkFolderList()
-        val mutableFoldersStateFlow =
-            bufferedMutableSharedFlow<List<SyncResponseJson.Folder>>(replay = 1)
-        every {
-            vaultDiskSource.getFolders(userId = MOCK_USER_STATE.activeUserId)
-        } returns mutableFoldersStateFlow
-        coEvery {
-            vaultSdkSource.decryptFolderList(
-                userId = userId,
-                folderList = mockEncryptedFolderList,
-            )
-        } returns throwable.asFailure()
-
-        vaultRepository
-            .foldersStateFlow
-            .test {
-                assertEquals(DataState.Loading, awaitItem())
-                mutableFoldersStateFlow.tryEmit(mockFolderList)
-
-                // No additional emissions until vault is unlocked
-                expectNoEvents()
-                setVaultToUnlocked(userId = userId)
-
-                assertEquals(DataState.Error<List<FolderView>>(throwable), awaitItem())
-            }
-    }
-
-    @Test
-    fun `sendDataStateFlow should emit decrypted list of sends when decryptSendsList succeeds`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockSendList = listOf(createMockSend(number = 1))
-            val mockEncryptedSendList = mockSendList.toEncryptedSdkSendList()
-            val mockSendViewList = listOf(createMockSendView(number = 1))
-            val mutableSendsStateFlow =
-                bufferedMutableSharedFlow<List<SyncResponseJson.Send>>(replay = 1)
-            every {
-                vaultDiskSource.getSends(userId = MOCK_USER_STATE.activeUserId)
-            } returns mutableSendsStateFlow
-            coEvery {
-                vaultSdkSource.decryptSendList(
-                    userId = userId,
-                    sendList = mockEncryptedSendList,
-                )
-            } returns mockSendViewList.asSuccess()
-
-            vaultRepository
-                .sendDataStateFlow
-                .test {
-                    assertEquals(DataState.Loading, awaitItem())
-                    mutableSendsStateFlow.tryEmit(mockSendList)
-
-                    // No additional emissions until vault is unlocked
-                    expectNoEvents()
-                    setVaultToUnlocked(userId = userId)
-
-                    assertEquals(DataState.Loaded(SendData(mockSendViewList)), awaitItem())
-                }
-        }
-
-    @Test
-    fun `sendDataStateFlow should emit an error when decryptSendsList fails`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        val userId = "mockId-1"
-        val throwable = Throwable("Fail")
-        val mockSendList = listOf(createMockSend(number = 1))
-        val mockEncryptedSendList = mockSendList.toEncryptedSdkSendList()
-        val mutableSendsStateFlow =
-            bufferedMutableSharedFlow<List<SyncResponseJson.Send>>(replay = 1)
-        every {
-            vaultDiskSource.getSends(userId = MOCK_USER_STATE.activeUserId)
-        } returns mutableSendsStateFlow
-        coEvery {
-            vaultSdkSource.decryptSendList(
-                userId = userId,
-                sendList = mockEncryptedSendList,
-            )
-        } returns throwable.asFailure()
-
-        vaultRepository
-            .sendDataStateFlow
-            .test {
-                assertEquals(DataState.Loading, awaitItem())
-                mutableSendsStateFlow.tryEmit(mockSendList)
-
-                // No additional emissions until vault is unlocked
-                expectNoEvents()
-                setVaultToUnlocked(userId = userId)
-
-                assertEquals(DataState.Error<SendData>(throwable), awaitItem())
-            }
     }
 
     @Test
@@ -738,238 +152,10 @@ class VaultRepositoryTest {
     }
 
     @Test
-    fun `databaseSchemeChangeFlow should trigger sync on emission`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        coEvery { vaultSyncManager.sync(any(), any()) } just awaits
-
-        mutableDatabaseSchemeChangeFlow.tryEmit(Unit)
-
-        coVerify(exactly = 1) { vaultSyncManager.sync(any(), any()) }
-    }
-
-    @Test
-    fun `sync should update DataStateFlow with an Error when vaultSyncManager result is Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val mockException = IllegalStateException("sad")
-            coEvery {
-                vaultSyncManager.sync(
-                    userId = MOCK_USER_STATE.activeUserId,
-                    forced = false,
-                )
-            } returns SyncVaultDataResult.Error(throwable = mockException)
-
-            vaultRepository.sync()
-
-            assertEquals(
-                DataState.Error<DecryptCipherListResult>(mockException),
-                vaultRepository.decryptCipherListResultStateFlow.value,
-            )
-            assertEquals(
-                DataState.Error<List<CollectionView>>(mockException),
-                vaultRepository.collectionsStateFlow.value,
-            )
-            assertEquals(
-                DataState.Error<DomainsData>(mockException),
-                vaultRepository.domainsStateFlow.value,
-            )
-            assertEquals(
-                DataState.Error<List<FolderView>>(mockException),
-                vaultRepository.foldersStateFlow.value,
-            )
-            assertEquals(
-                DataState.Error<SendData>(mockException),
-                vaultRepository.sendDataStateFlow.value,
-            )
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `sync should update vaultDataStateFlow with an Error when vaultSyncManager result is Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val mockException = IllegalStateException("sad")
-            coEvery {
-                vaultSyncManager.sync(any(), any())
-            } returns SyncVaultDataResult.Error(mockException)
-            setupVaultDiskSourceFlows()
-
-            vaultRepository
-                .vaultDataStateFlow
-                .test {
-                    assertEquals(DataState.Loading, awaitItem())
-                    vaultRepository.sync()
-                    assertEquals(DataState.Error<VaultData>(mockException), awaitItem())
-                }
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `sync should update DataStateFlows to NoNetwork when vaultSyncManager result is Error with `() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery {
-                vaultSyncManager.sync(any(), any())
-            } returns SyncVaultDataResult.Error(throwable = UnknownHostException())
-
-            vaultRepository.sync()
-
-            assertEquals(
-                DataState.NoNetwork(data = null),
-                vaultRepository.decryptCipherListResultStateFlow.value,
-            )
-            assertEquals(
-                DataState.NoNetwork(data = null),
-                vaultRepository.collectionsStateFlow.value,
-            )
-            assertEquals(
-                DataState.NoNetwork(data = null),
-                vaultRepository.domainsStateFlow.value,
-            )
-            assertEquals(
-                DataState.NoNetwork(data = null),
-                vaultRepository.foldersStateFlow.value,
-            )
-            assertEquals(
-                DataState.NoNetwork(data = null),
-                vaultRepository.sendDataStateFlow.value,
-            )
-        }
-
-    @Test
-    fun `sync with NoNetwork should update vaultDataStateFlow to NoNetwork`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        coEvery {
-            vaultSyncManager.sync(any(), any())
-        } returns SyncVaultDataResult.Error(UnknownHostException())
-        setupVaultDiskSourceFlows()
-
-        vaultRepository
-            .vaultDataStateFlow
-            .test {
-                assertEquals(DataState.Loading, awaitItem())
-                vaultRepository.sync()
-                assertEquals(DataState.NoNetwork(data = null), awaitItem())
-            }
-    }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `sync with NoNetwork data should update sendDataStateFlow to Pending and NoNetwork with data`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            setVaultToUnlocked(userId = userId)
-            coEvery {
-                vaultSyncManager.sync(
-                    userId = userId,
-                    forced = false,
-                )
-            } returns SyncVaultDataResult.Error(throwable = UnknownHostException())
-            val sendsFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Send>>()
-            setupVaultDiskSourceFlows(sendsFlow = sendsFlow)
-            coEvery {
-                vaultSdkSource.decryptSendList(
-                    userId = userId,
-                    sendList = listOf(createMockSdkSend(1)),
-                )
-            } returns listOf(createMockSendView(1)).asSuccess()
-
-            vaultRepository
-                .sendDataStateFlow
-                .test {
-                    assertEquals(DataState.Loading, awaitItem())
-                    sendsFlow.tryEmit(listOf(createMockSend(1)))
-                    assertEquals(
-                        DataState.Loaded(data = SendData(listOf(createMockSendView(1)))),
-                        awaitItem(),
-                    )
-                    vaultRepository.sync()
-                    assertEquals(
-                        DataState.Pending(data = SendData(listOf(createMockSendView(1)))),
-                        awaitItem(),
-                    )
-                    assertEquals(
-                        DataState.NoNetwork(data = SendData(listOf(createMockSendView(1)))),
-                        awaitItem(),
-                    )
-                }
-        }
-
-    @Test
-    fun `syncIfNecessary when there is no last sync time should sync the vault`() {
-        val userId = "mockId-1"
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        every {
-            settingsDiskSource.getLastSyncTime(userId = userId)
-        } returns null
-        coEvery { vaultSyncManager.sync(userId = userId, forced = false) } just awaits
-
-        vaultRepository.syncIfNecessary()
-
-        coVerify { vaultSyncManager.sync(userId = userId, forced = false) }
-    }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `syncIfNecessary when the current time is greater than 30 minutes after the last sync time should sync the vault`() {
-        val userId = "mockId-1"
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        every {
-            settingsDiskSource.getLastSyncTime(userId = userId)
-        } returns clock.instant().minus(31, ChronoUnit.MINUTES)
-        coEvery { vaultSyncManager.sync(userId = userId, forced = false) } just awaits
-
-        vaultRepository.syncIfNecessary()
-
-        coVerify { vaultSyncManager.sync(userId = userId, forced = false) }
-    }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `syncIfNecessary when the current time is less than 30 minutes after the last sync time should not sync the vault`() {
-        val userId = "mockId-1"
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        every {
-            settingsDiskSource.getLastSyncTime(userId = userId)
-        } returns clock.instant().minus(29, ChronoUnit.MINUTES)
-        coEvery { vaultSyncManager.sync(userId = userId, forced = false) } just awaits
-
-        vaultRepository.syncIfNecessary()
-
-        coVerify(exactly = 0) { vaultSyncManager.sync(userId = any(), forced = any()) }
-    }
-
-    @Test
-    fun `sync when the last sync time is older than the revision date should sync the vault`() {
-        val userId = "mockId-1"
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        every {
-            settingsDiskSource.getLastSyncTime(userId = userId)
-        } returns clock.instant().minus(1, ChronoUnit.MINUTES)
-
-        coEvery { vaultSyncManager.sync(userId = userId, forced = false) } just awaits
-
-        vaultRepository.sync()
-
-        coVerify { vaultSyncManager.sync(userId = userId, forced = false) }
-    }
-
-    @Test
-    fun `lockVaultForCurrentUser should delegate to the VaultLockManager`() {
-        vaultRepository.lockVaultForCurrentUser(isUserInitiated = true)
-        verify { vaultLockManager.lockVaultForCurrentUser(isUserInitiated = true) }
-    }
-
-    @Test
     fun `unlockVaultWithBiometrics with missing user state should return InvalidStateError`() =
         runTest {
             fakeAuthDiskSource.userState = null
             val cipher = mockk<Cipher>()
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
 
             val result = vaultRepository.unlockVaultWithBiometrics(cipher = cipher)
 
@@ -977,19 +163,11 @@ class VaultRepositoryTest {
                 VaultUnlockResult.InvalidStateError(error = NoActiveUserException()),
                 result,
             )
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
         }
 
     @Test
     fun `unlockVaultWithBiometrics with missing biometrics key should return InvalidStateError`() =
         runTest {
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
             fakeAuthDiskSource.userState = MOCK_USER_STATE
             val cipher = mockk<Cipher>()
             val userId = MOCK_USER_STATE.activeUserId
@@ -1000,10 +178,6 @@ class VaultRepositoryTest {
             assertEquals(
                 VaultUnlockResult.InvalidStateError(error = MissingPropertyException("Foo")),
                 result,
-            )
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
             )
         }
 
@@ -1375,10 +549,6 @@ class VaultRepositoryTest {
     fun `unlockVaultWithMasterPassword with missing user state should return InvalidStateError`() =
         runTest {
             fakeAuthDiskSource.userState = null
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
 
             val result = vaultRepository.unlockVaultWithMasterPassword(masterPassword = "")
 
@@ -1386,20 +556,11 @@ class VaultRepositoryTest {
                 VaultUnlockResult.InvalidStateError(error = NoActiveUserException()),
                 result,
             )
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
         }
 
     @Test
     fun `unlockVaultWithMasterPassword with missing user key should return InvalidStateError`() =
         runTest {
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
-
             val result = vaultRepository.unlockVaultWithMasterPassword(masterPassword = "")
             fakeAuthDiskSource.storeUserKey(
                 userId = "mockId-1",
@@ -1414,20 +575,12 @@ class VaultRepositoryTest {
                 VaultUnlockResult.InvalidStateError(error = MissingPropertyException("Foo")),
                 result,
             )
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
         }
 
     @Suppress("MaxLineLength")
     @Test
     fun `unlockVaultWithMasterPassword with missing private key should return InvalidStateError`() =
         runTest {
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
             val result = vaultRepository.unlockVaultWithMasterPassword(masterPassword = "")
             fakeAuthDiskSource.storeUserKey(
                 userId = "mockId-1",
@@ -1441,11 +594,6 @@ class VaultRepositoryTest {
             assertEquals(
                 VaultUnlockResult.InvalidStateError(error = MissingPropertyException("Foo")),
                 result,
-            )
-
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
             )
         }
 
@@ -1596,10 +744,6 @@ class VaultRepositoryTest {
     @Test
     fun `unlockVaultWithPin with missing user state should return InvalidStateError`() = runTest {
         fakeAuthDiskSource.userState = null
-        assertEquals(
-            emptyList<VaultUnlockData>(),
-            vaultRepository.vaultUnlockDataStateFlow.value,
-        )
 
         val result = vaultRepository.unlockVaultWithPin(pin = "1234")
 
@@ -1607,21 +751,12 @@ class VaultRepositoryTest {
             VaultUnlockResult.InvalidStateError(error = NoActiveUserException()),
             result,
         )
-        assertEquals(
-            emptyList<VaultUnlockData>(),
-            vaultRepository.vaultUnlockDataStateFlow.value,
-        )
     }
 
     @Suppress("MaxLineLength")
     @Test
     fun `unlockVaultWithPin with missing pin-protected user key should return InvalidStateError`() =
         runTest {
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
-
             val result = vaultRepository.unlockVaultWithPin(pin = "1234")
             fakeAuthDiskSource.storePinProtectedUserKey(
                 userId = "mockId-1",
@@ -1636,18 +771,10 @@ class VaultRepositoryTest {
                 VaultUnlockResult.InvalidStateError(error = MissingPropertyException("Foo")),
                 result,
             )
-            assertEquals(
-                emptyList<VaultUnlockData>(),
-                vaultRepository.vaultUnlockDataStateFlow.value,
-            )
         }
 
     @Test
     fun `unlockVaultWithPin with missing private key should return InvalidStateError`() = runTest {
-        assertEquals(
-            emptyList<VaultUnlockData>(),
-            vaultRepository.vaultUnlockDataStateFlow.value,
-        )
         val result = vaultRepository.unlockVaultWithPin(pin = "1234")
         fakeAuthDiskSource.storePinProtectedUserKey(
             userId = "mockId-1",
@@ -1661,10 +788,6 @@ class VaultRepositoryTest {
         assertEquals(
             VaultUnlockResult.InvalidStateError(error = MissingPropertyException("Foo")),
             result,
-        )
-        assertEquals(
-            emptyList<VaultUnlockData>(),
-            vaultRepository.vaultUnlockDataStateFlow.value,
         )
     }
 
@@ -1731,199 +854,60 @@ class VaultRepositoryTest {
         }
 
     @Test
-    fun `getVaultItemStateFlow should update to Error when a sync fails generically`() =
+    fun `getVaultItemStateFlow should update to Error when error state is emitted`() =
         runTest {
             val folderId = 1234
             val folderIdString = "mockId-$folderId"
             val throwable = Throwable("Fail")
             fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery {
-                vaultSyncManager.sync(any(), any())
-            } returns SyncVaultDataResult.Error(throwable)
-            setupVaultDiskSourceFlows()
 
             vaultRepository.getVaultItemStateFlow(folderIdString).test {
                 assertEquals(DataState.Loading, awaitItem())
-                vaultRepository.sync()
+                mutableVaultDataStateFlow.value = DataState.Error(throwable)
                 assertEquals(DataState.Error<CipherView>(throwable), awaitItem())
-            }
-
-            coVerify(exactly = 1) {
-                vaultSyncManager.sync(any(), any())
             }
         }
 
     @Test
-    fun `getVaultItemStateFlow should update to NoNetwork when a sync fails from no network`() =
+    fun `getVaultItemStateFlow should update to NoNetwork when a NoNetwork value is emitted`() =
         runTest {
             val itemId = 1234
             val itemIdString = "mockId-$itemId"
             fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery {
-                vaultSyncManager.sync(any(), any())
-            } returns SyncVaultDataResult.Error(UnknownHostException())
-            setupVaultDiskSourceFlows()
 
             vaultRepository.getVaultItemStateFlow(itemIdString).test {
                 assertEquals(DataState.Loading, awaitItem())
-                vaultRepository.sync()
+                mutableVaultDataStateFlow.value = DataState.NoNetwork()
                 assertEquals(DataState.NoNetwork<CipherView>(), awaitItem())
             }
-
-            coVerify(exactly = 1) {
-                vaultSyncManager.sync(any(), any())
-            }
         }
 
     @Test
-    fun `vaultDataStateFlow should return empty when last sync time is populated`() =
-        runTest {
-            val userId = "mockId-1"
-            coEvery {
-                vaultLockManager.waitUntilUnlocked(userId = userId)
-            } just runs
-            every {
-                settingsDiskSource.getLastSyncTime(userId = userId)
-            } returns clock.instant()
-
-            mutableVaultStateFlow.update {
-                listOf(VaultUnlockData(userId, VaultUnlockData.Status.UNLOCKED))
-            }
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            setupEmptyDecryptionResults()
-            setupVaultDiskSourceFlows(
-                ciphersFlow = flowOf(emptyList()),
-                collectionsFlow = flowOf(emptyList()),
-                domainsFlow = flowOf(
-                    SyncResponseJson.Domains(
-                        globalEquivalentDomains = emptyList(),
-                        equivalentDomains = emptyList(),
-                    ),
-                ),
-                foldersFlow = flowOf(emptyList()),
-                sendsFlow = flowOf(emptyList()),
-            )
-
-            turbineScope {
-                val ciphersStateFlow =
-                    vaultRepository.decryptCipherListResultStateFlow.testIn(backgroundScope)
-                val collectionsStateFlow =
-                    vaultRepository.collectionsStateFlow.testIn(backgroundScope)
-                val foldersStateFlow = vaultRepository.foldersStateFlow.testIn(backgroundScope)
-                val sendsStateFlow = vaultRepository.sendDataStateFlow.testIn(backgroundScope)
-                val domainsStateFlow = vaultRepository.domainsStateFlow.testIn(backgroundScope)
-
-                assertEquals(
-                    DataState.Loaded(
-                        createMockDecryptCipherListResult(
-                            number = 1,
-                            successes = emptyList(),
-                        ),
-                    ),
-                    ciphersStateFlow.awaitItem(),
-                )
-                assertEquals(
-                    DataState.Loaded(emptyList<CollectionView>()),
-                    collectionsStateFlow.awaitItem(),
-                )
-                assertEquals(
-                    DataState.Loaded(emptyList<FolderView>()),
-                    foldersStateFlow.awaitItem(),
-                )
-                assertEquals(
-                    DataState.Loaded(SendData(sendViewList = emptyList())),
-                    sendsStateFlow.awaitItem(),
-                )
-                assertEquals(
-                    DataState.Loaded(
-                        DomainsData(
-                            equivalentDomains = emptyList(),
-                            globalEquivalentDomains = emptyList(),
-                        ),
-                    ),
-                    domainsStateFlow.awaitItem(),
-                )
-            }
-        }
-
-    @Test
-    fun `vaultDataStateFlow should return loading when last sync time is null`() =
-        runTest {
-            val userId = "mockId-1"
-            coEvery {
-                vaultLockManager.waitUntilUnlocked(userId = userId)
-            } just runs
-            every {
-                settingsDiskSource.getLastSyncTime(userId = userId)
-            } returns null
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            setupEmptyDecryptionResults()
-            setupVaultDiskSourceFlows(
-                ciphersFlow = flowOf(emptyList()),
-                collectionsFlow = flowOf(emptyList()),
-                domainsFlow = flowOf(),
-                foldersFlow = flowOf(emptyList()),
-                sendsFlow = flowOf(emptyList()),
-            )
-            turbineScope {
-                val ciphersStateFlow =
-                    vaultRepository.decryptCipherListResultStateFlow.testIn(backgroundScope)
-                val collectionsStateFlow =
-                    vaultRepository.collectionsStateFlow.testIn(backgroundScope)
-                val foldersStateFlow = vaultRepository.foldersStateFlow.testIn(backgroundScope)
-                val sendsStateFlow = vaultRepository.sendDataStateFlow.testIn(backgroundScope)
-                val domainsStateFlow = vaultRepository.domainsStateFlow.testIn(backgroundScope)
-
-                assertEquals(DataState.Loading, ciphersStateFlow.awaitItem())
-                assertEquals(DataState.Loading, collectionsStateFlow.awaitItem())
-                assertEquals(DataState.Loading, foldersStateFlow.awaitItem())
-                assertEquals(DataState.Loading, sendsStateFlow.awaitItem())
-                assertEquals(DataState.Loading, domainsStateFlow.awaitItem())
-            }
-        }
-
-    @Test
-    fun `getVaultFolderStateFlow should update to NoNetwork when a sync fails from no network`() =
+    fun `getVaultFolderStateFlow should update to NoNetwork when no network value is emitted`() =
         runTest {
             val folderId = 1234
             val folderIdString = "mockId-$folderId"
             fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery {
-                vaultSyncManager.sync(any(), any())
-            } returns SyncVaultDataResult.Error(UnknownHostException())
-            setupVaultDiskSourceFlows()
 
             vaultRepository.getVaultFolderStateFlow(folderIdString).test {
                 assertEquals(DataState.Loading, awaitItem())
-                vaultRepository.sync()
+                mutableVaultDataStateFlow.value = DataState.NoNetwork()
                 assertEquals(DataState.NoNetwork<FolderView>(), awaitItem())
-            }
-
-            coVerify(exactly = 1) {
-                vaultSyncManager.sync(any(), any())
             }
         }
 
     @Test
-    fun `getVaultFolderStateFlow should update to Error when a sync fails generically`() =
+    fun `getVaultFolderStateFlow should update to Error when an error is emitted`() =
         runTest {
             val folderId = 1234
             val folderIdString = "mockId-$folderId"
             val throwable = Throwable("Fail")
             fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery {
-                vaultSyncManager.sync(any(), any())
-            } returns SyncVaultDataResult.Error(throwable)
-            setupVaultDiskSourceFlows()
 
             vaultRepository.getVaultFolderStateFlow(folderIdString).test {
                 assertEquals(DataState.Loading, awaitItem())
-                vaultRepository.sync()
+                mutableVaultDataStateFlow.value = DataState.Error(throwable)
                 assertEquals(DataState.Error<FolderView>(throwable), awaitItem())
-            }
-
-            coVerify(exactly = 1) {
-                vaultSyncManager.sync(any(), any())
             }
         }
 
@@ -1932,77 +916,40 @@ class VaultRepositoryTest {
         val sendId = 1
         fakeAuthDiskSource.userState = MOCK_USER_STATE
         val sendView = createMockSendView(number = sendId)
-        coEvery {
-            vaultSdkSource.decryptSendList(
-                userId = MOCK_USER_STATE.activeUserId,
-                sendList = emptyList(),
-            )
-        } returns emptyList<SendView>().asSuccess()
-        coEvery {
-            vaultSdkSource.decryptSendList(
-                userId = MOCK_USER_STATE.activeUserId,
-                sendList = listOf(createMockSdkSend(number = sendId)),
-            )
-        } returns listOf(sendView).asSuccess()
-
-        val sendsFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Send>>()
-        setupVaultDiskSourceFlows(sendsFlow = sendsFlow)
 
         vaultRepository.getSendStateFlow("mockId-$sendId").test {
             assertEquals(DataState.Loading, awaitItem())
-            sendsFlow.tryEmit(emptyList())
-
-            // No additional emissions until vault is unlocked
-            expectNoEvents()
-            setVaultToUnlocked(userId = MOCK_USER_STATE.activeUserId)
-
-            sendsFlow.tryEmit(emptyList())
+            mutableSendDataStateFlow.value = DataState.Loaded(SendData(emptyList()))
             assertEquals(DataState.Loaded<SendView?>(null), awaitItem())
-            sendsFlow.tryEmit(listOf(createMockSend(number = sendId)))
+            mutableSendDataStateFlow.value = DataState.Loaded(SendData(listOf(sendView)))
             assertEquals(DataState.Loaded<SendView?>(sendView), awaitItem())
         }
     }
 
     @Test
-    fun `getSendStateFlow should update to NoNetwork when a sync fails from no network`() =
+    fun `getSendStateFlow should update to NoNetwork when NoNetwork value is emitted`() =
         runTest {
             val sendId = 1234
             fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery {
-                vaultSyncManager.sync(any(), any())
-            } returns SyncVaultDataResult.Error(UnknownHostException())
-            setupVaultDiskSourceFlows()
 
             vaultRepository.getSendStateFlow("mockId-$sendId").test {
                 assertEquals(DataState.Loading, awaitItem())
-                vaultRepository.sync()
+                mutableSendDataStateFlow.value = DataState.NoNetwork()
                 assertEquals(DataState.NoNetwork<SendView?>(), awaitItem())
-            }
-
-            coVerify(exactly = 1) {
-                vaultSyncManager.sync(any(), any())
             }
         }
 
     @Test
-    fun `getSendStateFlow should update to Error when a sync fails generically`() =
+    fun `getSendStateFlow should update to Error when an error is emitted`() =
         runTest {
             val sendId = 1234
             val throwable = Throwable("Fail")
             fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery {
-                vaultSyncManager.sync(any(), any())
-            } returns SyncVaultDataResult.Error(throwable)
-            setupVaultDiskSourceFlows()
 
             vaultRepository.getSendStateFlow("mockId-$sendId").test {
                 assertEquals(DataState.Loading, awaitItem())
-                vaultRepository.sync()
+                mutableSendDataStateFlow.value = DataState.Error(throwable)
                 assertEquals(DataState.Error<SendView?>(throwable), awaitItem())
-            }
-
-            coVerify(exactly = 1) {
-                vaultSyncManager.sync(any(), any())
             }
         }
 
@@ -2025,13 +972,18 @@ class VaultRepositoryTest {
     @Test
     fun `generateTotp should return a success result on getting a code`() = runTest {
         val totpResponse = TotpResponse("Testcode", 30u)
-        val userId = "mockId-1"
         coEvery {
             vaultSdkSource.generateTotpForCipherListView(any(), any(), any())
         } returns totpResponse.asSuccess()
+        mutableDecryptCipherListResultStateFlow.value = DataState.Loaded(
+            DecryptCipherListResult(
+                successes = listOf(
+                    createMockCipherListView(number = 1),
+                ),
+                failures = emptyList(),
+            ),
+        )
         fakeAuthDiskSource.userState = MOCK_USER_STATE
-        setVaultToUnlocked(userId = userId)
-        setupDataStateFlow(userId = userId)
 
         val result = vaultRepository.generateTotp(
             cipherId = "mockId-1",
@@ -2058,23 +1010,24 @@ class VaultRepositoryTest {
         runTest {
             fakeAuthDiskSource.userState = MOCK_USER_STATE
             val userId = "mockId-1"
-            setVaultToUnlocked(userId = userId)
-
-            coEvery {
-                vaultSyncManager.sync(userId = userId, forced = false)
-            } returns SyncVaultDataResult.Success(itemsAvailable = true)
-
-            val stateFlow = MutableStateFlow<DataState<VerificationCodeItem?>>(
-                DataState.Loading,
-            )
+            val stateFlow = MutableStateFlow<DataState<VerificationCodeItem?>>(DataState.Loading)
 
             every {
                 totpCodeManager.getTotpCodeStateFlow(userId = userId, cipherListView = any())
             } returns stateFlow
+            mutableVaultDataStateFlow.value = DataState.Loaded(
+                data = VaultData(
+                    decryptCipherListResult = DecryptCipherListResult(
+                        successes = listOf(createMockCipherListView(number = 1)),
+                        failures = emptyList(),
+                    ),
+                    collectionViewList = emptyList(),
+                    folderViewList = emptyList(),
+                    sendViewList = emptyList(),
+                ),
+            )
 
-            setupDataStateFlow(userId = userId)
-
-            vaultRepository.getAuthCodeFlow(createMockCipherView(1).id.toString()).test {
+            vaultRepository.getAuthCodeFlow(userId).test {
                 assertEquals(
                     DataState.Loading,
                     awaitItem(),
@@ -2084,13 +1037,6 @@ class VaultRepositoryTest {
 
                 assertEquals(
                     DataState.Loaded(createVerificationCodeItem()),
-                    awaitItem(),
-                )
-
-                vaultRepository.sync()
-
-                assertEquals(
-                    DataState.Pending(createVerificationCodeItem()),
                     awaitItem(),
                 )
             }
@@ -2106,15 +1052,7 @@ class VaultRepositoryTest {
     fun `getAuthCodesFlow should update data state when state changes`() = runTest {
         fakeAuthDiskSource.userState = MOCK_USER_STATE
         val userId = "mockId-1"
-        setVaultToUnlocked(userId = userId)
-
-        coEvery {
-            vaultSyncManager.sync(any(), any())
-        } returns SyncVaultDataResult.Success(itemsAvailable = true)
-
-        val stateFlow = MutableStateFlow<DataState<List<VerificationCodeItem>>>(
-            DataState.Loading,
-        )
+        val stateFlow = MutableStateFlow<DataState<List<VerificationCodeItem>>>(DataState.Loading)
 
         every {
             totpCodeManager.getTotpCodesForCipherListViewsStateFlow(
@@ -2122,8 +1060,17 @@ class VaultRepositoryTest {
                 cipherListViews = any(),
             )
         } returns stateFlow
-
-        setupDataStateFlow(userId = userId)
+        mutableVaultDataStateFlow.value = DataState.Loaded(
+            data = VaultData(
+                decryptCipherListResult = DecryptCipherListResult(
+                    successes = listOf(createMockCipherListView(number = 1)),
+                    failures = emptyList(),
+                ),
+                collectionViewList = emptyList(),
+                folderViewList = emptyList(),
+                sendViewList = emptyList(),
+            ),
+        )
 
         vaultRepository.getAuthCodesFlow().test {
             assertEquals(
@@ -2137,28 +1084,7 @@ class VaultRepositoryTest {
                 DataState.Loaded(listOf(createVerificationCodeItem())),
                 awaitItem(),
             )
-
-            vaultRepository.sync()
-
-            assertEquals(
-                DataState.Pending(listOf(createVerificationCodeItem())),
-                awaitItem(),
-            )
         }
-    }
-
-    @Test
-    fun `fullSyncFlow emission should trigger unforced sync`() {
-        val userId = "mockId-1"
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        every {
-            settingsDiskSource.getLastSyncTime(userId = userId)
-        } returns null
-        coEvery { vaultSyncManager.sync(any(), any()) } just awaits
-
-        mutableFullSyncFlow.tryEmit(Unit)
-
-        coVerify { vaultSyncManager.sync(any(), any()) }
     }
 
     @Suppress("MaxLineLength")
@@ -2293,9 +1219,8 @@ class VaultRepositoryTest {
             )
         }
 
-    @Suppress("MaxLineLength")
     @Test
-    fun `importCxfPayload should return success result when payload is successfully imported and vault sync is successful`() =
+    fun `importCxfPayload should return success result when payload is successfully imported`() =
         runTest {
             val userId = "mockId-1"
             val payload = "payload"
@@ -2307,18 +1232,12 @@ class VaultRepositoryTest {
                     payload = payload,
                 )
             } returns ImportCxfPayloadResult.Success(itemCount = 1)
-            coEvery {
-                vaultSyncManager.sync(userId = userId, forced = true)
-            } returns SyncVaultDataResult.Success(itemsAvailable = true)
             val result = vaultRepository.importCxfPayload(payload)
 
             assertEquals(
                 ImportCredentialsResult.Success(itemCount = 1),
                 result,
             )
-            coVerify(exactly = 1) {
-                vaultSyncManager.sync(userId = userId, forced = true)
-            }
         }
 
     @Test
@@ -2349,9 +1268,6 @@ class VaultRepositoryTest {
             ImportCredentialsResult.Error(expected),
             result,
         )
-        coVerify(exactly = 0) {
-            vaultSyncManager.sync(userId = userId, forced = true)
-        }
     }
 
     @Test
@@ -2373,9 +1289,6 @@ class VaultRepositoryTest {
             ImportCredentialsResult.NoItems,
             result,
         )
-        coVerify(exactly = 0) {
-            vaultSyncManager.sync(userId = userId, forced = true)
-        }
     }
 
     @Test
@@ -2390,11 +1303,7 @@ class VaultRepositoryTest {
                 userId = userId,
                 payload = payload,
             )
-        } returns ImportCxfPayloadResult.Success(itemCount = 1)
-
-        coEvery {
-            vaultSyncManager.sync(userId = userId, forced = true)
-        } returns SyncVaultDataResult.Error(throwable)
+        } returns ImportCxfPayloadResult.SyncFailed(throwable)
 
         val result = vaultRepository.importCxfPayload(payload)
 
@@ -2491,51 +1400,6 @@ class VaultRepositoryTest {
         }
     }
 
-    @Suppress("MaxLineLength")
-    @Test
-    fun `syncForResult should return success result with itemsAvailable = true when sync succeeds and ciphers list is not empty`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            coEvery {
-                vaultSyncManager.sync(
-                    userId = userId,
-                    forced = false,
-                )
-            } returns SyncVaultDataResult.Success(itemsAvailable = true)
-
-            val syncResult = vaultRepository.syncForResult()
-            assertEquals(SyncVaultDataResult.Success(itemsAvailable = true), syncResult)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `syncForResult should return success result with itemsAvailable = false when sync succeeds and ciphers list is empty`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            coEvery {
-                vaultSyncManager.sync(userId = userId, forced = false)
-            } returns SyncVaultDataResult.Success(itemsAvailable = false)
-
-            val syncResult = vaultRepository.syncForResult()
-            assertEquals(SyncVaultDataResult.Success(itemsAvailable = false), syncResult)
-        }
-
-    @Test
-    fun `syncForResult should return error when VaultSyncManager sync result is Error`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-        val throwable = Throwable()
-        coEvery {
-            vaultSyncManager.sync(any(), any())
-        } returns SyncVaultDataResult.Error(throwable)
-        val syncResult = vaultRepository.syncForResult()
-        assertEquals(
-            SyncVaultDataResult.Error(throwable = throwable),
-            syncResult,
-        )
-    }
-
     //region Helper functions
 
     /**
@@ -2548,7 +1412,7 @@ class VaultRepositoryTest {
     ) {
         val userId = "mockId-1"
         coEvery {
-            vaultSyncManager.sync(any(), any())
+            vaultSyncManager.syncForResult(forced = any())
         } returns SyncVaultDataResult.Success(itemsAvailable = true)
         coEvery {
             vaultSdkSource.decryptSendList(
@@ -2608,128 +1472,6 @@ class VaultRepositoryTest {
             )
         } returns unlockResult
     }
-
-    /**
-     * Ensures the vault for the given [userId] is unlocked and can pass any
-     * [VaultLockManager.waitUntilUnlocked] or [VaultLockManager.isVaultUnlocked] checks.
-     */
-    private fun setVaultToUnlocked(userId: String) {
-        mutableUnlockedUserIdsStateFlow.update { it + userId }
-        mutableVaultStateFlow.tryEmit(
-            listOf(
-                VaultUnlockData(
-                    userId,
-                    VaultUnlockData.Status.UNLOCKED,
-                ),
-            ),
-        )
-    }
-
-    /**
-     * Helper setup all flows required to properly subscribe to the
-     * [VaultRepository.vaultDataStateFlow].
-     */
-    private fun setupVaultDiskSourceFlows(
-        ciphersFlow: Flow<List<SyncResponseJson.Cipher>> = bufferedMutableSharedFlow(),
-        collectionsFlow: Flow<List<SyncResponseJson.Collection>> = bufferedMutableSharedFlow(),
-        domainsFlow: Flow<SyncResponseJson.Domains> = bufferedMutableSharedFlow(),
-        foldersFlow: Flow<List<SyncResponseJson.Folder>> = bufferedMutableSharedFlow(),
-        sendsFlow: Flow<List<SyncResponseJson.Send>> = bufferedMutableSharedFlow(),
-    ) {
-        coEvery { vaultDiskSource.getCiphersFlow(MOCK_USER_STATE.activeUserId) } returns ciphersFlow
-        coEvery {
-            vaultDiskSource.getCollections(MOCK_USER_STATE.activeUserId)
-        } returns collectionsFlow
-        coEvery { vaultDiskSource.getDomains(MOCK_USER_STATE.activeUserId) } returns domainsFlow
-        coEvery { vaultDiskSource.getFolders(MOCK_USER_STATE.activeUserId) } returns foldersFlow
-        coEvery { vaultDiskSource.getSends(MOCK_USER_STATE.activeUserId) } returns sendsFlow
-    }
-
-    private fun setupEmptyDecryptionResults() {
-        coEvery {
-            vaultSdkSource.decryptCipherListWithFailures(
-                userId = MOCK_USER_STATE.activeUserId,
-                cipherList = emptyList(),
-            )
-        } returns createMockDecryptCipherListResult(number = 1, successes = emptyList()).asSuccess()
-        coEvery {
-            vaultSdkSource.decryptFolderList(
-                userId = MOCK_USER_STATE.activeUserId,
-                folderList = emptyList(),
-            )
-        } returns emptyList<FolderView>().asSuccess()
-        coEvery {
-            vaultSdkSource.decryptCollectionList(
-                userId = MOCK_USER_STATE.activeUserId,
-                collectionList = emptyList(),
-            )
-        } returns emptyList<CollectionView>().asSuccess()
-
-        coEvery {
-            vaultSdkSource.decryptSendList(
-                userId = MOCK_USER_STATE.activeUserId,
-                sendList = emptyList(),
-            )
-        } returns emptyList<SendView>().asSuccess()
-    }
-
-    private suspend fun setupDataStateFlow(userId: String) {
-        coEvery {
-            vaultSdkSource.decryptCipherListWithFailures(
-                userId = userId,
-                cipherList = listOf(createMockSdkCipher(1, clock)),
-            )
-        } returns createMockDecryptCipherListResult(number = 1).asSuccess()
-        coEvery {
-            vaultSdkSource.decryptFolderList(
-                userId = userId,
-                folderList = listOf(createMockSdkFolder(1)),
-            )
-        } returns listOf(createMockFolderView(number = 1)).asSuccess()
-        coEvery {
-            vaultSdkSource.decryptCollectionList(
-                userId = userId,
-                collectionList = listOf(createMockSdkCollection(1)),
-            )
-        } returns listOf(createMockCollectionView(number = 1)).asSuccess()
-        coEvery {
-            vaultSdkSource.decryptSendList(
-                userId = userId,
-                sendList = listOf(createMockSdkSend(number = 1)),
-            )
-        } returns listOf(createMockSendView(number = 1)).asSuccess()
-        val ciphersFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Cipher>>()
-        val collectionsFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Collection>>()
-        val foldersFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Folder>>()
-        val sendsFlow = bufferedMutableSharedFlow<List<SyncResponseJson.Send>>()
-        setupVaultDiskSourceFlows(
-            ciphersFlow = ciphersFlow,
-            collectionsFlow = collectionsFlow,
-            foldersFlow = foldersFlow,
-            sendsFlow = sendsFlow,
-        )
-
-        vaultRepository.vaultDataStateFlow.test {
-            ciphersFlow.tryEmit(listOf(createMockCipher(1)))
-            collectionsFlow.tryEmit(listOf(createMockCollection(number = 1)))
-            foldersFlow.tryEmit(listOf(createMockFolder(number = 1)))
-            sendsFlow.tryEmit(listOf(createMockSend(number = 1)))
-            assertEquals(DataState.Loading, awaitItem())
-
-            assertEquals(
-                DataState.Loaded(
-                    data = VaultData(
-                        decryptCipherListResult = createMockDecryptCipherListResult(number = 1),
-                        collectionViewList = listOf(createMockCollectionView(number = 1)),
-                        folderViewList = listOf(createMockFolderView(number = 1)),
-                        sendViewList = listOf(createMockSendView(number = 1)),
-                    ),
-                ),
-                awaitItem(),
-            )
-        }
-    }
-
     //endregion Helper functions
 }
